refactor: split AIArchitect service into smaller components
- Created shared types in types.ts
- Split PromptManager into its own class
- Created OpenAIService for API interactions
- Updated main AIArchitect to use new components
- Fixed TypeScript configuration
